Zoom and offset can be set globally; Activitys of the same kind can not overlap anymore; An activity startpoint can not be after the end anymore; start and end can be set via keyboard or popupmenu (popup automatically selects the activity depending on cursor); ScrollPane for visualizations;

+++ EvalTool/src/evaluationtool/DataModel.java
@@ -71,7 +71,7 @@ public class DataModel {
 	 */
 	public void setPlaybackPosition(float pos, boolean setVideoPosition) {
 		for(int i = 0; i < loadedDataTracks.size(); i++){
-			  loadedDataTracks.get(i).getVisualization().setPosition(pos * gui.getVideoLength());
+			  loadedDataTracks.get(i).getVisualization().setPosition((long)(pos * gui.getVideoLength()));
 		  }
 		if(setVideoPosition)
 			gui.setVideoPosition(pos);
